# public/api-docs/models/openapi.yaml
openapi: 3.1.1
info:
  title: '🏛️ TDW - ACiencia REST api'
  description: '[UPM] TDW REST api ACiencia'
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.8.0
  contact:
    name: Máster Universitario en Ingeniería Web
    url: 'https://miw.etsisi.upm.es/'
    email: miw.etsisi@upm.es
servers:
  - url: '/api/v1'
    description: TDW Production server (uses live data)
...
paths:
...
  '/asociaciones':
    get:
      tags:
        - '🏛 Asociaciones'
      summary: Retrieves the collection of Asociacion resources.
      description: Returns all asociaciones from the system.
      operationId: readAsociaciones
      parameters:
        - $ref: '#/components/parameters/elementTextParam'
        - $ref: '#/components/parameters/orderParam'
        - $ref: '#/components/parameters/orderingParam'
      responses:
        '200':
          description: Object containing an array of Asociaciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  asociaciones:
                    description: Asociaciones array
                    type: array
                    items:
                      $ref: '#/components/schemas/Asociacion'
        '404':
          $ref: '#/components/responses/Response_404'
    post:
      tags:
        - '🏛 Asociaciones'
      summary: Creates an Asociacion resource.
      description: Creates a new asociacion.
      operationId: createAsociacion
      security:
        - MiWApiSecurity: [ writer ]
      requestBody:
            $ref: '#/components/requestBodies/AsociacionCreateRequestBody'
      responses:
        '201':
          description: '`Created`: asociacion created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asociacion'
          headers:
            Location:
              $ref: '#/components/headers/Location'
        '400':
          $ref: '#/components/responses/Response_400'
        '401':
          $ref: '#/components/responses/Response_401'
        '403':
          $ref: '#/components/responses/Response_403'
        '422':
          $ref: '#/components/responses/Response_422'
    options:
      tags:
        - '🏛 Asociaciones'
      summary: Provides the list of HTTP supported methods.
      description: Return an `Allow` header with a comma separated list of HTTP supported methods.
      operationId: optionsAsociaciones
      responses:
        '204':
          description: '`Allow` header <Response body is empty>'
          headers:
            Allow:
              $ref: '#/components/headers/Allow'
...
components:
  schemas:
...
    Element:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Element Id
          readOnly: true
        name:
          type: string
          maxLength: 80
        birthDate:
          description: An RFC3339 formatted date string ("2025-01-01")
          type: string
          format: date
        deathDate:
          description: An RFC3339 formatted date string ("2025-01-01")
          type: string
          format: date
        imageUrl:
          type: string
          maxLength: 2047
          format: url
        wikiUrl:
          type: string
          maxLength: 2047
          format: url
      required:
        - name
...
    Asociacion:
      allOf:
        - $ref: '#/components/schemas/Element'
        - type: object
          properties:
            websiteUrl:
              description: URL del sitio web oficial de la asociación
              type: string
              format: uri
            entities:
              description: Array of Entity Ids
              type: array
              items:
                type: integer
    Message:
      type: object
      description: HTTP response message
      properties:
        code:
          description: Response code
          type: integer
          format: int64
        message:
          description: Response message
          type: string
  responses:
...
    Response_400:
      description: '`BAD REQUEST`: User name or e-mail already exists, or role does not exist'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-400:
              $ref: '#/components/examples/response-400'
    Response_401:
      description: '`UNAUTHORIZED`: invalid Authorization header'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-401:
              $ref: '#/components/examples/response-401'
    Response_403:
      description: '`FORBIDDEN`: You don''t have permission to access'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-403:
              $ref: '#/components/examples/response-403'
    Response_404:
      description: '`NOT FOUND`: resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404:
              $ref: '#/components/examples/response-404'
...
    Response_422:
      description: '`UNPROCESSABLE ENTITY`: Username, e-mail or password is left out'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-422:
              $ref: '#/components/examples/response-422'
  requestBodies:
...
    AsociacionCreateRequestBody:
      description: '`Asociacion` data to create'
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                description: Name of the asociacion
              websiteUrl:
                type: string
                description: Website URL of the asociacion
              birthDate:
                type: string
                format: date
                description: Birth date (optional)
              deathDate:
                type: string
                format: date
                description: Death date (optional)
              imageUrl:
                type: string
                description: URL to an image (optional)
              wikiUrl:
                type: string
                description: Wikipedia URL (optional)
            required:
              - name
              - websiteUrl
  parameters:
...
    elementTextParam:
      name: name
      in: query
      description: Elements containing the text
      required: false
      schema:
        type: string
        pattern: '^[a-zA-Z0-9()áéíóúÁÉÍÓÚñÑ %$\.+-<>]+$'
...
    orderParam:
      name: order
      in: query
      description: '`id` | `name`'
      required: false
      allowEmptyValue: false
      schema:
        type: string
        enum:
          - id
          - name
    orderingParam:
      name: ordering
      in: query
      description: '`ASC` | `DESC`'
      required: false
      allowEmptyValue: false
      schema:
        type: string
        enum:
          - ASC
          - DESC
...
  headers:
...
    Location:
      description: '`URI` of the new resource'
      schema:
        type: string
        format: url
    Allow:
      description: List of HTTP supported methods
      schema:
        type: string
  examples:
    response-400:
      value:
        code: 400
        message: 'BAD REQUEST: name or e-mail already exists'
    response-401:
      value:
        code: 401
        message: 'UNAUTHORIZED: invalid Authorization header'
    response-403:
      value:
        code: 403
        message: "FORBIDDEN: You don't have permission to access"
    response-404:
      value:
        code: 404
        message: "NOT FOUND: Resource not found (or you don't have permission to access)"
...
    response-422:
      value:
        code: 422
        message: 'UNPROCESSABLE ENTITY: name, e-mail or password is left out'
  securitySchemes:
    MiWApiSecurity:
      type: oauth2
      description: 'This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)'
      bearerFormat: JWT
      flows:
        password:
          tokenUrl: /access_token
          scopes:
            reader: 'Permissions: (read, update) [ User, Person, Entity, Product ]'
            writer: 'can also: (create, update, delete) [ User, Person, Entity, Product ]'